+++ src/domains/matches/matches-history/mocks/matches-history-handler.ts
@@ -5,8 +5,7 @@ import MatchesHistoryDummy from "@matches/matches-history/mocks/matches-history-
 
 export const MatchesHistoryHandlers = [
   // Supabase RPC URL 패턴을 더 정확하게 가로채기
-  http.post("*/rest/v1/rpc/get_all_finish_match_lists", ({ request }) => {
-    console.log("MSW: RPC 요청 가로챔!", request.url);
+  http.post("*/rest/v1/rpc/get_all_finish_match_lists", () => {
     return HttpResponse.json(MatchesHistoryDummy);
   }),
 ];