+++ src/domains/players/players-rating-rotator/mocks/player-rating-rotator-handlers.ts
@@ -4,8 +4,7 @@ import { HttpResponse, http } from "msw";
 
 export const PlayersRatingRotatorHandlers = [
   // Supabase RPC URL 패턴을 더 정확하게 가로채기
-  http.post("*/rest/v1/rpc/get_player_rating_rotator_acc", ({ request }) => {
-    console.log("MSW: RPC 요청 가로챔!", request.url);
+  http.post("*/rest/v1/rpc/get_player_rating_rotator_acc", () => {
     return HttpResponse.json(PlayersRatingRotatorDummy);
   }),
 ];